Cleanup of tile URL retrieval for aperture rest calls. Tracking of data in the client layer, so it only requests stuff it doesn't already have, and stores all it has been given.  Next step: all() and rendering

+++ tile-service/src/main/java/com/oculusinfo/tile/rest/ImageTileResource.java
@@ -106,9 +106,9 @@ public class ImageTileResource extends ApertureServerResource {
 			String levelDir = (String) getRequest().getAttributes().get("level");
 			int zoomLevel = Integer.parseInt(levelDir);
 			String xAttr = (String) getRequest().getAttributes().get("x");
-			double x = Double.parseDouble(xAttr);
+			int x = Integer.parseInt(xAttr);
 			String yAttr = (String) getRequest().getAttributes().get("y");
-			double y = Double.parseDouble(yAttr);
+			int y = Integer.parseInt(yAttr);
 
 			String ext = (String) getRequest().getAttributes().get("ext");
 			ExtensionType extType = ExtensionType.valueOf(ext.trim().toLowerCase());
